[feature] Contact, Account in InvitationService.java : Delete 기능 구현

+++ src/main/java/jun/invitation/domain/gallery/Service/GalleryService.java
@@ -24,6 +24,9 @@ public class GalleryService {
 
     @Transactional
     public void delete(List<Gallery> galleries) {
+        if (galleries.isEmpty()) {
+            return;
+        }
         galleries.forEach(g -> {
             s3UploadService.delete(g.getStoreFileName());
         });
